Add files via upload
The following is an IEEE formatted article dissecting the analytical findings from a Maryland car sales data set. This article will include images of graphs displaying the relationship between used car sales and new sales, followed by how well the sales were trajected to perform in 2023. I used Microsoft Excel to perform the analysis and Tableau to create the visualizations.

Figure 6 was created using Tableau; the visual shows that used cars being forecasted to surpass the sales of new cars.

The following findings can be impactful for dealerships to use for specific sales to meet quota.